Factor out AssertingSearcher so it can be used by mock Engines
